requirements.txt
# Core ML libraries
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
xgboost>=2.0.0

# MLflow and experiment tracking
mlflow>=2.9.0

# Vetiver and Pins
vetiver>=0.2.0
pins>=0.8.0

# API framework
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.4.0

# Utilities
pyyaml>=6.0
python-dotenv>=1.0.0
joblib>=1.3.0
pyarrow>=21.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# Code quality
black>=23.9.0
flake8>=6.1.0
mypy>=1.5.0

# model deployment
huggingface-hub